Clean up scripts. Remove useless prefab.
