--- Assets/QA/Input_Test/Prefabs/Standard_Gamepad.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 4aa7549d08cee4cac9c3e59ead230907
-timeCreated: 1509652010
-licenseType: Pro
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
