--- Assets/QA/Input_Test/Prefabs/Xbox_Gamepad.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: cf58c9049f9c94f1e95f27ab4d900fd5
-timeCreated: 1509997557
-licenseType: Pro
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
