--- Assets/QA/Input_Test/Scenes/Controller Xbox.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: 8f5a18c7b34093e4b8e68821516bb499
-DefaultImporter:
-  userData:
