--- Assets/QA/Input_Test/Scenes/Mouse Keyboard.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: 401d9a8859c068f4db3331d02e2854a1
-DefaultImporter:
-  userData:
