--- Assets/QA/Input_Test/Scripts/ClickOnDropdown.cs
@@ -1,19 +1,0 @@
-using UnityEngine;
-
-public class ClickOnDropdown : MonoBehaviour
-{
-    public delegate void OnClickDropdownDelegate();
-    public static OnClickDropdownDelegate clickDropdown;
-
-    //public void OnPointerClick(PointerEventData eventData)
-    //{
-    //    if (eventData.button == PointerEventData.InputButton.Left)
-    //        clickDropdown();
-    //}
-
-    private void Update()
-    {
-        if (Input.GetMouseButtonUp(0))
-            clickDropdown();
-    }
-}
