+++ Assets/QA/Input_Test/Scripts/Input/ControllerDiagramForInputSystem.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 8e2c41f95d8884994a805dd8e700df74
-timeCreated: 1516325340
-licenseType: Pro
+guid: 5a00ddc04a6f4a94885e88952bbc534e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
