+++ Assets/QA/Input_Test/Scripts/Input/XboxForInputSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: abd144405b2a59444af6a0e1f2c06ea4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
